requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
passlib==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3

